feat(security): client-safe @revealui/security barrel via /server split
Makes the default @revealui/security barrel client-bundle-safe - the durable
structural fix for the node:crypto client-crash class (the richtext-RSC crash
was point-fixed earlier; this removes the footgun at the source).

- New @revealui/security/server subpath re-exports the node:-bearing modules:
  auth/gdpr/audit (node:crypto), ssrf (node:dns, incl. createSafeFetch +
  SafeFetchOptions). The barrel re-exports only client-safe modules (alerting,
  authorization, encryption [Web Crypto], gdpr-storage, headers, logger,
  request-ip, sanitize). dist/index.js is crypto-free; dist/server.js carries
  the node: graph.
- @revealui/core/security re-exports BOTH barrel + /server, so consumers via
  core/security are unchanged. Direct-barrel consumers of the moved symbols are
  routed to /server: auth audit-bridge, mcp adapter + remote-client, server
  marketplace route, admin MCP-OAuth initiate route (the last three are #1034's
  ssrf consumers). Test mocks retargeted to '@revealui/security/server'.
- apps/server/vitest.config.ts: add explicit /server + /sanitize source aliases
  before the bare '@revealui/security' alias (Vite's prefix-match would
  otherwise rewrite the subpath to '.../src/index.ts/server' -> ENOTDIR).
- validate:client-safety gate: server-only set is now /server + core/security;
  the barrel is an enforced client-safe entry; flags STATIC node:/server only
  (dynamic await import() is code-split + bundle-safe, e.g. audit.ts).

Verified on rebased test: barrel dist crypto-free; typecheck 24/24; gate PASS;
@revealui/security 690 + mcp remote-client 11 + admin oauth 11 + server
marketplace 82 tests pass; biome clean. changeset: @revealui/security minor +
@revealui/core patch.
