+++ packages/mcp/__tests__/remote-client.test.ts
@@ -24,9 +24,9 @@ vi.mock('../src/client.js', async () => {
 // the hermetic test environment (no network) before the McpClient is ever
 // constructed. These tests exercise the options→McpClient plumbing, not the
 // guard — the guard itself is covered by @revealui/security's ssrf.test.ts.
-// remote-client.ts is the only consumer of @revealui/security in this test's
-// module graph, and it imports just these two symbols.
-vi.mock('@revealui/security', () => ({
+// remote-client.ts is the only consumer of @revealui/security/server in this
+// test's module graph, and it imports just these two symbols.
+vi.mock('@revealui/security/server', () => ({
   assertPublicUrl: vi.fn().mockResolvedValue(undefined),
   createSafeFetch: vi.fn(() => globalThis.fetch),
 }));
